Extract request from controller and use it in service
