+++ web-services-app/src/main/java/edu/unc/lib/boxc/web/services/processing/RunEnhancementsService.java
@@ -57,12 +57,13 @@ public class RunEnhancementsService {
      * Service to take a list of pids searches for file objects which are in the list of pids
      * or children of those objects and run enhancements on.
      *
-     * @param agent security principals of the agent making request.
-     * @param objectPids List of pids to run enhancements on
-     * @param force whether enhancements should run if derivatives are already present
+     * @param request Request to run enhancements
      */
-    public void run(AgentPrincipals agent, List<String> objectPids, boolean force) {
+    public void run(RunEnhancementsRequest request) {
         try (Timer.Context context = timer.time()) {
+            var agent = request.getAgent();
+            var objectPids = request.getPids();
+            var force = request.isForce();
             for (String objectPid : objectPids) {
                 PID pid = PIDs.get(objectPid);
